requirements.txt
aiohttp
protobuf
pycryptodome
python-cfonts
protobuf-decoder
urllib3
flask
gunicorn